+++ src/struct/graph.h
@@ -5,4 +5,4 @@
 struct datum **parse_inf_graph(struct datum **idlist, const char *fn,
                                char *src, char *dest);
 
-uint32 **init_graph(struct datum *src, struct datum **idlist);
+uint64 **init_graph(struct datum *src, struct datum **idlist);
